move test to parent class. move resolver back to JdbcSqlConnector [HZ-2340]
